Module Private.Parser

Attempts to parse a buffer into a HTTP request. If successful, it returns the parsed request and an offset value that indicates the starting point of unconsumed content left in the buffer.

type error =

| Partial

| Msg of string

val parse_request : ?pos:int -> ?len:int -> string -> (Request.t * int, error) Stdlib.result

val parse_chunk_length : ?pos:int -> ?len:int -> string -> (int64 * int, error) Stdlib.result